minor refactor updates to unit test

+++ libraries/predictors/tcc/SingleInputThresholdRule.tcc
@@ -9,7 +9,7 @@
 namespace predictors
 {
     template<typename RandomAccessVectorType>
-    size_t predictors::SingleInputThresholdRule::Compute(const RandomAccessVectorType& inputVector) const
+    int predictors::SingleInputThresholdRule::Compute(const RandomAccessVectorType& inputVector) const
     {
         return inputVector[_index] > _threshold ? 1 : 0;
     }
